Wifi: separate lwip-related code into files
